Make bootstrap-compatible
I found that this CSS rule for the checkbox was overridden by the bootstrap.css, and the width declaration in particular reverted to 'auto'.  By adding the type selector, this overrides the bootstrap CSS, and the switch responds to clicks over its entire width, not just the leftmost few pixels.